# swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Altzone API",
        "version": "1.0.0",
        "description": "This is a description of API for the Altzone game.\nThis swagger document is describing endpoint in more details and includes the request objects structure, examples, authorization rules and any other additional information.\nNotice,  that it does not include common rules and functions, since they are covered on the wikipages in GitHub",
        "x-logo": {
            "url": ""
        }
    },
    "servers": [
        {
            "url": "https://altzone.fi/api",
            "description": "Production. Stable version of API, which might not have all of the new changes"
        },
        {
            "url": "http://localhost:8080",
            "description": "Local"
        },
        {
            "url": "https://devapi.altzone.fi",
            "description": "Development. Current version of API with latest changes"
        }
    ],
    "paths": {
...
        "/item/{_id}": {
            "get": {
                "tags": [
                    "Item"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/200_one"
                                },
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "Item": {
                                                    "_id": "668a70ce91020196cb10d595",
                                                    "name": "Sofa_Taakka",
                                                    "recycling": "Mixed_waste",
                                                    "qualityLevel": "common",
                                                    "unityKey": "Sofa_Taakka",
                                                    "weight": 67,
                                                    "location": [
                                                        2,
                                                        4
                                                    ],
                                                    "isFurniture": true,
                                                    "stock_id": "668953cd4a7b6c993aed1a36",
                                                    "price": 23
                                                }
                                            },
                                            "metaData": {
                                                "dataKey": "Item",
                                                "modelName": "Item",
                                                "dataType": "Object",
                                                "dataCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "$ref": "#/components/responses/400_New"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_New"
                    },
                    "500": {
                        "$ref": "#/components/responses/500_New"
                    }
                },
                "security": [
                    {}
                ],
                "summary": "Get Item by _id",
                "description": "Read Item data by its _id field\n\n| Feature / Requirement | Is / Has |\n| -------- | -------- |\n| Authentication | no |\n| Authorization   | no |\n\n-------------\n\nThe name field is an enum:\n\n|                |               |                 |              |\n|------------------------|-----------------------|--------------------------|-----------------------|\n| 'Sofa_Taakka'          | 'Mirror_Taakka'       | 'Floorlamp_Taakka'        | 'Toilet_Schrodinger'  |\n| 'Sink_Schrodinger'     | 'Closet_Taakka'       | 'SofaTable_Taakka'        | 'SideTable_Taakka'    |\n| 'Carpet_Schrodinger'   | 'Carpet_Rakkaus'      | 'Mirror_Schrodinger'      | 'Mirror_Rakkaus'      |\n| 'ArmChair_Taakka'      | 'Sofa_Rakkaus'        | 'ArmChair_Rakkaus'        | 'Closet_Rakkaus'      |\n| 'Chair_Neuro'          | 'Dresser_Neuro'       | 'Stool_Neuro'             |                       |\n\n-------------\n\nThe recycling field is an enum:\n\n|               |              |    |           |\n|-----------------------|-----------------------|------------|-------------------|\n| 'Mixed_waste'         | 'Electrical_Equipment'| 'Wood'     | 'Glass'           |\n| 'Metal'               | 'Landfill'            |            |                   |\n\n-------------\n\nThe qualityLevel field is an enum:\n\n|    |   |    |    |\n|------------|------------|------------|------------|\n| 'common'   | 'rare'     | 'epic'     |            |\n\n-------------\n\n## Endpoint description\n\n### General description\n\nItems are objects, which can be placed in SoulHomes' Rooms, Stocks and Shop. Item is an object, which has its value in the game and can be bought from the Shop, stolen from the Stock or placed in safe to SoulHome's Room.\n\nItems can be obtained only from Shop and then stored in Room or a Stock.\n\nItem can be a furniture or some other kind of object. The difference between furniture and not furniture is that a furniture can be placed only on a floor. Not furniture can be placed anywhere including on top of a furniture, walls, celling and floor. This data is used in game logic only and does not affect the API in any way.\n\n### Relations\n\nItem can belong to one Room. Room can have multiple Items.\n\nItem can belong to one Stock. Stock can have multiple Items.\n\nItem can belong to Shop. Shop can have multiple Items.\n\n### CRUD operations\n\nItem can not be created by Player. It can only be bought from Shop. API creates new Items automatically.\n\nItem can be read by anyone. One Item can be accessed via its _id field. Multiple Items can be accessed by requesting them within another endpoints via \"with\" or \"all\" queries (/stock/:stock_id?with=Item), since they are mostly required together with Stock, Room or Shop data.\n\nItem can not be directly updated by Players. However, Players may move Items from one place to another. They first can buy an Item from Shop and so update Item's place by moving it to the Clan's Stock. Then they can move it from SoulHome to Stock or vice versa. Item can be bought from Shop via /stock/buy enpoint (not yet implemented). Item can be moved from Room to Stock or from Stock to Room via /item/move POST endpoint.\n\nItem can not be deleted."
            },
            "parameters": [
                {
                    "examples": {
                        "1": {
                            "value": "668a70ce91020196cb10d595"
                        }
                    },
                    "name": "_id",
                    "description": "item _id field",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
        "/item/steal": {
            "get": {
                "tags": [
                    "Item"
                ],
                "parameters": [
                    {
                        "name": "steal_token",
                        "description": "",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "SoulHome": {
                                                    "_id": "668c0e9f9470819cf38e371c",
                                                    "name": "clan1",
                                                    "clan_id": "667bfec6afb8211b4bd8dbff",
                                                    "Room": [
                                                        {
                                                            "_id": "668d6003a9292e300e94c833",
                                                            "floorType": "floor1",
                                                            "wallType": "wall1",
                                                            "isActive": false,
                                                            "deactivationTimestamp": 34524523654,
                                                            "cellCount": 20,
                                                            "soulHome_id": "668c0e9f9470819cf38e371c",
                                                            "hasLift": true
                                                        }
                                                    ]
                                                }
                                            },
                                            "metaData": {
                                                "dataKey": "SoulHome",
                                                "modelName": "SoulHome",
                                                "dataType": "Object",
                                                "dataCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "SoulHome and its Rooms"
                    },
                    "400": {
                        "$ref": "#/components/responses/400_New"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_New"
                    },
                    "403": {
                        "$ref": "#/components/responses/403_New"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_New"
                    },
                    "500": {
                        "$ref": "#/components/responses/500_New"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Get SoulHome data from which items can be stolen",
                "description": "Based on the provided steal token, which contains the SoulHome _id from which Items can be stolen, the SoulHome data and its Rooms will be returned.\n\nYou can see the process flow from [this diagram](https://github.com/Alt-Org/Altzone-Server/tree/dev/doc/img/game_results)"
            },
            "post": {
                "requestBody": {
                    "description": "Items to be stolen",
                    "content": {
                        "application/json": {
                            "examples": {
                                "1": {
                                    "value": {
                                        "steal_token": "some_token",
                                        "item_ids": [
                                            "668a765091020196cb10d5a3",
                                            "668a765091020196cb10d5a3"
                                        ],
                                        "room_id": "668a765091020196cb10d5a3"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Item"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "Item": [
                                                    {
                                                        "_id": "668a70ce91020196cb10d595",
                                                        "name": "item1",
                                                        "weight": 67,
                                                        "recycling": "yleinen",
                                                        "unityKey": "unity1",
                                                        "filename": "file1",
                                                        "location": [
                                                            2,
                                                            4
                                                        ],
                                                        "isFurniture": true,
                                                        "stock_id": "668953cd4a7b6c993aed1a36",
                                                        "price": 23
                                                    }
                                                ]
                                            },
                                            "metaData": {
                                                "dataKey": "Item",
                                                "modelName": "Item",
                                                "dataType": "Array",
                                                "dataCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Array of Items, which were stolen"
                    },
                    "400": {
                        "$ref": "#/components/responses/400_New"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_New"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_New"
                    },
                    "500": {
                        "$ref": "#/components/responses/500_New"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Steal Items from another Clan",
                "description": "Steal Items from the loser Clan's SoulHome. The stolen Items will be automatically added to a specified Room of the winners Clan's SoulHome.\n\nNotice that at first a steal token should be obtained by the winner player(s) from the /gameData/battle POST endpoint with body type \"result\", while informing about game result.\n\nRequests without the steal token or with an expired token will get 403 as a response.\n\nNotice that only found SoulHome Items will be stolen. It is possible that some other player already has stoled some of the specified Items, in this case these Items will be ignored since they can not be stolen twice.  \n\nYou can see the process flow from [this diagram](https://github.com/Alt-Org/Altzone-Server/tree/dev/doc/img/game_results)\n\n## Endpoint description\n\n### General description\n\nItems are objects, which can be placed in SoulHomes' Rooms, Stocks and Shop. Item is an object, which has its value in the game and can be bought from the Shop, stolen from the Stock or placed in safe to SoulHome's Room.\n\nItems can be obtained only from Shop and then stored in Room or a Stock.\n\nItem can be a furniture or some other kind of object. The difference between furniture and not furniture is that a furniture can be placed only on a floor. Not furniture can be placed anywhere including on top of a furniture, walls, celling and floor. This data is used in game logic only and does not affect the API in any way.\n\n### Relations\n\nItem can belong to one Room. Room can have multiple Items.\n\nItem can belong to one Stock. Stock can have multiple Items.\n\nItem can belong to Shop. Shop can have multiple Items.\n\n### CRUD operations\n\nItem can not be created by Player. It can only be bought from Shop. API creates new Items automatically.\n\nItem can be read by anyone. One Item can be accessed via its _id field. Multiple Items can be accessed by requesting them within another endpoints via \"with\" or \"all\" queries (/stock/:stock_id?with=Item), since they are mostly required together with Stock, Room or Shop data.\n\nItem can not be directly updated by Players. However, Players may move Items from one place to another. They first can buy an Item from Shop and so update Item's place by moving it to the Clan's Stock. Then they can move it from SoulHome to Stock or vice versa. Item can be bought from Shop via /stock/buy enpoint (not yet implemented). Item can be moved from Room to Stock or from Stock to Room via /item/move POST endpoint.\n\nItem can not be deleted."
            }
        },
...
        "/item/move": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "1": {
                                    "value": {
                                        "item_id": "397bfec6adb8211b4bd8dbe0",
                                        "moveTo": "Stock",
                                        "destination_id": "667bfec6afb8211b4bd8dbff"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Item"
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "400": {
                        "$ref": "#/components/responses/400_New"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_New"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_New"
                    },
                    "500": {
                        "$ref": "#/components/responses/500_New"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Move Item",
                "description": "Move Item from Stock to Room or from Room to Stock, based on the _moveTo_ field, which can have only two values: \"Stock\" or \"Room\"\n\nNotice that Clan members can move only own Clan Items and a member is trying to move other Clan's Item the 404 will be returned.\n\nNotice that if an Item need to be moved to a Stock then there is no need to specify the destination_id field, since Clan can have only one Stock.\n\n| Feature / Requirement | Is / Has |\n| -------- | -------- |\n| Authentication | yes |\n| Authorization   | no |\n\n## Endpoint description\n\n### General description\n\nItems are objects, which can be placed in SoulHomes' Rooms, Stocks and Shop. Item is an object, which has its value in the game and can be bought from the Shop, stolen from the Stock or placed in safe to SoulHome's Room.\n\nItems can be obtained only from Shop and then stored in Room or a Stock.\n\nItem can be a furniture or some other kind of object. The difference between furniture and not furniture is that a furniture can be placed only on a floor. Not furniture can be placed anywhere including on top of a furniture, walls, celling and floor. This data is used in game logic only and does not affect the API in any way.\n\n### Relations\n\nItem can belong to one Room. Room can have multiple Items.\n\nItem can belong to one Stock. Stock can have multiple Items.\n\nItem can belong to Shop. Shop can have multiple Items.\n\n### CRUD operations\n\nItem can not be created by Player. It can only be bought from Shop. API creates new Items automatically.\n\nItem can be read by anyone. One Item can be accessed via its _id field. Multiple Items can be accessed by requesting them within another endpoints via \"with\" or \"all\" queries (/stock/:stock_id?with=Item), since they are mostly required together with Stock, Room or Shop data.\n\nItem can not be directly updated by Players. However, Players may move Items from one place to another. They first can buy an Item from Shop and so update Item's place by moving it to the Clan's Stock. Then they can move it from SoulHome to Stock or vice versa. Item can be bought from Shop via /stock/buy enpoint (not yet implemented). Item can be moved from Room to Stock or from Stock to Room via /item/move POST endpoint.\n\nItem can not be deleted."
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "200_one": {
                "description": "Found object. The response objects has two fields: data and metaData.\nThe data has the requested data. It has an object inside, which key is a collection name.\n\nThe metaData contains additional data, which can be used for parsing the data object: \ndataKey - collection name, in data object,\ndataType - type of object in data object.",
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "metaData": {
                        "type": "object",
                        "properties": {
                            "dataKey": {
                                "description": "collection name",
                                "type": "string"
                            },
                            "modelName": {
                                "description": "model name",
                                "type": "string"
                            },
                            "dataType": {
                                "description": "always has Object value",
                                "type": "string"
                            },
                            "dataCount": {
                                "description": "always has 1 value",
                                "type": "integer"
                            }
                        }
                    }
                }
            },
...
            "ErrorValidation": {
                "title": "Error on Validation",
                "description": "Error happen, because some of the fields are not valid",
                "type": "object",
                "properties": {
                    "response": {
                        "type": "string",
                        "example": "NOT_STRING"
                    },
                    "status": {
                        "format": "int32",
                        "description": "HTTP status code of the error",
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "nullable": true,
                        "description": "Error message explaining the error reason",
                        "type": "string",
                        "example": "name must be a string"
                    },
                    "name": {
                        "description": "Name of the error",
                        "type": "string",
                        "example": ""
                    },
                    "reason": {
                        "description": "Why the error happen",
                        "enum": [
                            "BAD_REQUEST",
                            "REQUIRED",
                            "NOT_ALLOWED",
                            "VALIDATION",
                            "NOT_STRING",
                            "NOT_NUMBER",
                            "NOT_BOOLEAN",
                            "NOT_ARRAY",
                            "NOT_OBJECT",
                            "NOT_DATE",
                            "WRONG_ENUM",
                            "LESS_THAN_MIN",
                            "MORE_THAN_MAX"
                        ],
                        "type": "string",
                        "example": "NOT_STRING"
                    },
                    "field": {
                        "nullable": true,
                        "description": "On which field the error happen",
                        "type": "string",
                        "example": "name"
                    },
                    "value": {
                        "nullable": true,
                        "type": "string",
                        "example": 1
                    },
                    "additional": {
                        "nullable": true,
                        "description": "Additional data, which can be useful. For unexpected errors this field will include the error happen",
                        "type": "string",
                        "example": "isString"
                    },
                    "statusCode": {
                        "format": "int32",
                        "description": "HTTP status code of the error",
                        "type": "integer",
                        "example": 400
                    },
                    "objectType": {
                        "description": "Type of the object. All errors are of type APIError",
                        "type": "string",
                        "example": "APIError"
                    }
                }
            },
...
        },
        "responses": {
...
            "204": {
                "description": "Successful request, response with no body"
            },
...
            "400_New": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "format": "int32",
                                    "description": "HTTP status code of the response",
                                    "type": "integer",
                                    "example": 400
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ErrorValidation"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Validation error. Some of the fields are not specified, have wrong data types or any over validation problem"
            },
            "401_New": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "format": "int32",
                                    "description": "HTTP status code of the response",
                                    "type": "integer",
                                    "example": 401
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ErrorAuthentication"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Not authenticated. The Authorization header is not specified or the token is expired. [Here is more about auth](https://github.com/Alt-Org/Altzone-Server/wiki/2.-Authentication-and-authorization)"
            },
            "403_New": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "format": "int32",
                                    "description": "HTTP status code of the response",
                                    "type": "integer",
                                    "example": 403
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ErrorAuthorization"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "No permission. The logged-in user has no permission to execute the action. [Here is more about auth](https://github.com/Alt-Org/Altzone-Server/wiki/2.-Authentication-and-authorization)"
            },
            "404_New": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "format": "int32",
                                    "description": "HTTP status code of the response",
                                    "type": "integer",
                                    "example": 404
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ErrorNotFound"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Not found. No object(s) found"
            },
...
            "500_New": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode": {
                                    "format": "int32",
                                    "description": "HTTP status code of the response",
                                    "type": "integer",
                                    "example": 500
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ErrorServer"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Unexpected error happened. [Please create a new issue here](https://github.com/Alt-Org/Altzone-Server/issues) and specify the endpoint, HTTP method and description if u have any"
            }
        },
...
        "securitySchemes": {
            "JWTAuth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Example of header: \nAuthorization: Bearer access_token"
            }
        }
    },
    "security": [
        {
            "JWTAuth": []
        }
    ],
...
}